Update EECMT_cars.ini
Edited file to name all cars in the GOTY edition of Dirt Rally 2.0.  Requires replacement of the "VictorBush.Ego.Nefslib.dll" file with the version from NefsEDIT 0.4.0 for EECMT to find the extra 30 cars. Later versions won't work with EECMT, due to a change in the structure of the library file.  Also seems to need you to move all the folders out of the "cars" folder in Dirt Rally 2.0, to force EECMT to recreate all the camera files.  If you have individual car settings saved, move the folders out and then overwrite them back in afterwards, otherwise just delete them all if like me you just make a batch adjustment to all cars the same.